# docs/swagger.yaml
basePath: /am-fuel-gas-webapi
definitions:
  models.AuthUserData:
    properties:
      authStatus:
        type: integer
      displayName:
        type: string
      domain:
        type: string
      name:
        type: string
      permission:
        items:
          $ref: '#/definitions/models.MyPermission'
        type: array
      type:
        type: string
    type: object
  models.CalculationList:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.DensityCalculationHistory:
    properties:
      calculationDate:
        type: string
      endDate:
        type: string
      error:
        type: string
      startDate:
        type: string
      syncMode:
        type: string
      userName:
        type: string
      value:
        type: number
    type: object
  models.GetCalculatedImbalanceDetails:
    properties:
      imbalanceCalculation:
        $ref: '#/definitions/models.ImbalanceCalculation'
      nodes:
        items:
          $ref: '#/definitions/models.Node'
        type: array
    type: object
  models.GetDensityCoefficient:
    properties:
      calculationHistory:
        items:
          $ref: '#/definitions/models.DensityCalculationHistory'
        type: array
      densityCoefficient:
        type: number
    type: object
  models.GetManualFuelGas:
    properties:
      description:
        type: string
      id:
        type: integer
      lastUpdateDate:
        type: string
      name:
        type: string
      tag:
        type: string
      timestamp:
        type: string
      value:
        type: number
    type: object
  models.GetScales:
    properties:
      description:
        type: string
      id:
        type: integer
      value:
        type: string
    type: object
  models.ImbalanceCalculation:
    properties:
      aggregateTotal:
        type: string
      autoTotal:
        type: string
      calculationDate:
        type: string
      grp10Manual:
        type: string
      id:
        type: string
      manualTotal:
        type: string
      nitka1Auto:
        type: string
      nitka1Manual:
        type: string
      nitka2Auto:
        type: string
      nitka2Manual:
        type: string
      nitka3Auto:
        type: string
      nitka3Manual:
        type: string
      pgRedisTotal:
        type: string
      userName:
        type: string
    type: object
  models.ImbalanceCalculationHistory:
    properties:
      calculationDate:
        type: string
      endDate:
        type: string
      error:
        type: string
      id:
        type: string
      startDate:
        type: string
      syncMode:
        type: string
      userName:
        type: string
    type: object
  models.MyPermission:
    properties:
      name:
        type: string
      permission:
        type: integer
    type: object
  models.Node:
    properties:
      adjustment:
        type: string
      batchId:
        type: string
      consumption:
        type: string
      description:
        type: string
      distributed:
        type: string
      flag:
        type: string
      gasRedistribution:
        type: string
      id:
        type: integer
      value:
        type: string
    type: object
  models.NodeList:
    properties:
      description:
        type: string
      flag:
        type: string
      id:
        type: integer
    type: object
  models.SetImbalanceFlag:
    properties:
      adjustment:
        type: string
      flag:
        type: string
      id:
        type: integer
    type: object
  models.SetManualFuelGas:
    properties:
      date:
        type: string
      id:
        type: integer
      tag:
        type: string
      value:
        type: number
    type: object
  models.UpdateHistory:
    properties:
      timestampInsert:
        type: string
      value:
        type: number
    type: object
  models.UpdateScale:
    properties:
      id:
        type: integer
      value:
        type: string
    type: object
  models.UserData:
    properties:
      domain:
        type: string
      login:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/Authorization/GetCurrentUserInfo:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthUserData'
      tags:
      - Authorization
  /api/Authorization/LogInAuthorization:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: userdata
        required: true
        schema:
          $ref: '#/definitions/models.UserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthUserData'
      tags:
      - Authorization
  /api/Authorization/LogOutAuthorization:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Authorization
  /api/CalculateImbalance:
    post:
      consumes:
      - application/json
      parameters:
      - description: Дата получения параметров
        in: query
        name: date
        required: true
        type: string
      - description: Id batch расчета
        in: query
        name: batch
        required: true
        type: string
      - description: Данные расчета небаланс
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/models.SetImbalanceFlag'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Calculations
  /api/GetCalculatedImbalanceDetails:
    get:
      consumes:
      - application/json
      parameters:
      - description: Id batch расчета
        in: query
        name: batch
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetCalculatedImbalanceDetails'
      tags:
      - Calculations
  /api/GetCalculationsList:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CalculationList'
      tags:
      - Calculations
  /api/GetDensityCoefficientDetails:
    get:
      consumes:
      - application/json
      parameters:
      - description: Дата получения параметров
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetDensityCoefficient'
      tags:
      - Calculations
  /api/GetImbalanceHistory:
    get:
      consumes:
      - application/json
      parameters:
      - description: Дата получения параметров
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ImbalanceCalculationHistory'
            type: array
      tags:
      - Calculations
  /api/GetNodesList:
    get:
      consumes:
      - application/json
      parameters:
      - description: Id batch расчета
        in: query
        name: cloneId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NodeList'
      tags:
      - Calculations
  /api/GetParameterHistory:
    get:
      consumes:
      - application/json
      parameters:
      - description: Дата получения параметров
        in: query
        name: date
        required: true
        type: string
      - description: Id параметра
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateHistory'
      tags:
      - Parameters
  /api/GetParameters:
    get:
      consumes:
      - application/json
      parameters:
      - description: Дата получения параметров
        in: query
        name: date
        required: true
        type: string
      - description: 'Тэг типы: ''day'', ''month'' or empty'
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetManualFuelGas'
      tags:
      - Parameters
  /api/GetScales:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetScales'
      tags:
      - Parameters
  /api/PrepareImbalanceCalculation:
    post:
      consumes:
      - application/json
      parameters:
      - description: Дата расчета
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - Calculations
  /api/RecalculateDensityCoefficient:
    post:
      consumes:
      - application/json
      parameters:
      - description: Дата получения параметров
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Calculations
  /api/RemoveImbalanceCalculation:
    post:
      consumes:
      - application/json
      parameters:
      - description: Id batch расчета
        in: query
        name: batch
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Calculations
  /api/SetParameters:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные газ
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SetManualFuelGas'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Parameters
  /api/UpdateScale:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные шкалы
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateScale'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - Parameters
swagger: "2.0"